// src/components/types.ts
export interface FontProsp {
    color?: string;
    fontFamily?: string;
    font?: string;
    fontWeight?: string | number;
    lineHeight?: string;
    fontStyle?: string;
    paddingLeft?: string;
    paddingTop?: string;
    letterSpacing?: string;
    textAlign?: string;
    marginLeft?: string;
}

export interface WrapperProsp {
    width?: string;
    height?: string;
    margin?: string;
    paddingLeft?: string;
}